move implementation of the asset manager to its own module.
